app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai>=1.40.0
pydantic>=2.7.0
python-dotenv>=1.0.0
spotipy>=2.23.0
mcp>=1.0.0
